vboot: move timer and beep functions to vboot2 namespace
Move these functions from vboot_api.h into 2api.h:
  VbExGetTimer --> vb2ex_mtime (usec -> msec)
  VbExSleepMs --> vb2ex_msleep
  VbExBeep --> vb2ex_beep

Rename the constants:
  VB_USEC_PER_MSEC --> VB2_USEC_PER_MSEC
  VB_MSEC_PER_SEC --> VB2_MSEC_PER_SEC
Remove the constant VB_USEC_PER_SEC.

The error code VBERROR_NO_BACKGROUND_SOUND is dropped since
it is not currently used.

Update a few printf lines to use VB2_DEBUG instead.

BUG=b:124141368, chromium:988410
TEST=make clean && make runtests
BRANCH=none

Cq-Depend: chromium:2158665

+++ firmware/2lib/include/2ui.h
@@ -62,7 +62,7 @@ struct vb2_ui_context {
 
 	/* For developer mode screen. */
 	int disable_timer;
-	uint64_t start_time;
+	uint32_t start_time;
 	int beep_count;
 
 	/* For check_shutdown_request. */
